Group FieldView components and move canvas.js to /components
