+++ FtcDashboard/dash/src/components/ConfigurableLayout.tsx
@@ -8,7 +8,7 @@ import 'react-resizable/css/styles.css';
 
 import { ConfigurableView } from '../enums/ConfigurableView';
 import GraphView from '../containers/GraphView';
-import FieldView from '../containers/FieldView';
+import FieldView from '../containers/FieldView/FieldView';
 import ConfigView from '../containers/ConfigView';
 import TelemetryView from '../containers/TelemetryView';
 import CameraView from '../containers/CameraView';
