+++ FtcDashboard/dash/src/containers/FieldView/FieldView.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import BaseView, { BaseViewHeading } from './BaseView';
+import BaseView, { BaseViewHeading } from '../BaseView';
 import Field from './Field';
-import AutoFitCanvas from '../components/AutoFitCanvas';
+import AutoFitCanvas from '../../components/AutoFitCanvas';
 
 class FieldView extends React.Component {
   constructor(props) {
